fix for make run error
while deploying workbench, the `make run` command was giving an error and saying "unknown attribute quck_withdraw_limit for Currency" and deleting that line of code seems to be a fix for that problem. However the workbench is pulling 1.9.1 rc.22, so if the attribute is added in the later versions, updating the pulled release could also fix the problem.